:truck: move `get_wallet_id_from_b64encoded_jwt` from test utils to main utils
